actualizando cambios

+++ database/migrations/2018_07_30_000206_create_product_features_table.php
@@ -16,8 +16,8 @@ class CreateProductFeaturesTable extends Migration
         Schema::create('product_features', function (Blueprint $table) {
             $table->increments('id');
             $table->string('code', 10)->nullable()->default(null);
-            $table->double('price', 8, 2)->nullable()->default(null);
-            $table->float('iva', 5)->nullable()->default(null);
+            $table->double('price', 12, 2)->nullable()->default(null);
+            $table->double('iva', 12, 2)->nullable()->default(null);
 
             $table->unsignedInteger('product_feature_category_id')->default(1);
             $table->foreign('product_feature_category_id')->references('id')->on('product_feature_categories');
